Finaliza organização do plugin em vários arquivos
