+++ delibera_comments.php
@@ -1,8 +1,4 @@
 <?php
-require_once('delibera_comments_query.php');
-require_once('delibera_comments_template.php');
-require_once('delibera_comments_edit.php');
-
 
 function delibera_get_comment_type($comment)
 {
@@ -802,3 +798,147 @@ function delibera_comment_number_filtro($count, $postID)
 }
 
 add_filter('get_comments_number', 'delibera_comment_number_filtro', 10, 2);
+
+/**
+ * Sempre que um usuário valida uma pauta
+ * verifica se o número mínimo de validações foi
+ * atingido e se sim muda a situação da pauta de
+ * "emvotacao" para "discussao".
+ *
+ * @param unknown $post
+ * @return null
+ */
+function delibera_valida_validacoes($post)
+{
+	$validacoes = get_post_meta($post, 'numero_validacoes', true);
+	$min_validacoes = get_post_meta($post, 'min_validacoes', true);
+
+	if($validacoes >= $min_validacoes)
+	{
+		wp_set_object_terms($post, 'discussao', 'situacao', false); //Mudar situação para Discussão
+		if(has_action('delibera_validacao_concluida'))
+		{
+			do_action('delibera_validacao_concluida', $post);
+		}
+	}
+	else
+	{
+		if(has_action('delibera_validacao'))
+		{
+			do_action('delibera_validacao', $post);
+		}
+	}
+}
+
+/* Faz os testes de permissões para garantir que nenhum engraçadinho
+ * está injetando variáveis maliciosas.
+ * TODO: Incluir todas as variaveis a serem verificadas aqui
+ */
+function delibera_valida_permissoes($comment_ID)
+{
+	if (get_post_type() == 'pauta' && !delibera_current_user_can_participate())
+	{
+		if (array_key_exists('delibera_validacao', $_REQUEST) || array_key_exists('delibera_encaminha', $_REQUEST) )
+			wp_die("Nananina não! Você não tem que ter permissão pra votar.","Tocooo!!");
+	}
+}
+add_action( 'wp_blacklist_check', 'delibera_valida_permissoes' );
+
+/**
+ *
+ * Verifica se o número de votos é igual ao número de representantes para deflagar fim da votação
+ * @param integer $postID
+ */
+function delibera_valida_votos($postID)
+{
+	global $wp_roles,$wpdb;
+	$users_count = 0;
+    foreach ($wp_roles->roles as $nome => $role)
+    {
+    	if(is_array($role['capabilities']) && array_key_exists('votar', $role['capabilities']) && $role['capabilities']['votar'] == 1 ? "SSSSSim" : "NNNnnnnnnnao")
+    	{
+    		$result = $wpdb->get_results("SELECT count(*) as n FROM $wpdb->usermeta WHERE meta_key = 'wp_capabilities' AND meta_value LIKE '%$nome%' ");
+    		$users_count += $result[0]->n;
+    	}
+    }
+
+	$votos = delibera_get_comments_votacoes($postID);
+
+	$votos_count = count($votos);
+
+	if($votos_count >= $users_count)
+	{
+		delibera_computa_votos($postID, $votos);
+	}
+}
+
+/**
+ *
+ * Faz a apuração dos votos e toma as devidas ações:
+ *    Empate: Mais prazo;
+ *    Vencedor: Marco com resolucao e marca o encaminhamento.
+ * @param interger $postID
+ * @param array $votos
+ */
+function delibera_computa_votos($postID, $votos = null)
+{
+	if(is_null($votos)) // Ocorre no fim do prazo de votação
+	{
+		$votos = delibera_get_comments_votacoes($postID);
+	}
+	$encaminhamentos = delibera_get_comments_encaminhamentos($postID);
+	$encaminhamentos_votos = array();
+	foreach ($encaminhamentos as $encaminhamento)
+	{
+		$encaminhamentos_votos[$encaminhamento->comment_ID] = 0;
+	}
+
+	foreach ($votos as $voto_comment)
+	{
+		$voto = get_comment_meta($voto_comment->comment_ID, 'delibera_votos', true);
+		foreach ($voto as $voto_para)
+		{
+            if (isset($encaminhamentos_votos[$voto_para]))
+            {
+                $encaminhamentos_votos[$voto_para]++;
+            } else {
+                $encaminhamentos_votos[$voto_para] = 1;
+            }
+		}
+	}
+	$maisvotado = array(-1, -1);
+	$iguais = array();
+
+	foreach ($encaminhamentos_votos as $encaminhamentos_voto_key => $encaminhamentos_voto_valor)
+	{
+		if($encaminhamentos_voto_valor > $maisvotado[1])
+		{
+			$maisvotado[0] = $encaminhamentos_voto_key;
+			$maisvotado[1] = $encaminhamentos_voto_valor;
+			$iguais = array();
+		}
+		elseif($encaminhamentos_voto_valor == $maisvotado[1])
+		{
+			$iguais[] = $encaminhamentos_voto_key;
+		}
+		delete_comment_meta($encaminhamentos_voto_key, 'delibera_comment_numero_votos');
+		add_comment_meta($encaminhamentos_voto_key, 'delibera_comment_numero_votos', $encaminhamentos_voto_valor, true);
+	}
+
+	// nao finaliza a votacao caso haja um empate, exceto quando o administrador clicar no botão "Forçar fim do prazo"
+	if(count($iguais) > 0 && !(isset($_REQUEST['action']) && $_REQUEST['action'] == 'delibera_forca_fim_prazo_action')) // Empato
+	{
+		delibera_novo_prazo($postID);
+	}
+	else
+	{
+		wp_set_object_terms($postID, 'comresolucao', 'situacao', false);
+		update_comment_meta($maisvotado[0], 'delibera_comment_tipo', 'resolucao');
+		add_post_meta($postID, 'data_resolucao', date('d/m/Y H:i:s'), true);
+		////delibera_notificar_situacao($postID);
+		if(has_action('votacao_concluida'))
+		{
+			do_action('votacao_concluida', $post);
+		}
+	}
+}
